whatsapp-webhook/pyproject.toml
[project]
name = "whatsapp-webhook"
version = "0.1.0"
description = "WhatsApp bridge webhook"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-dialogflow-cx>=2.3.0",
    "flask>=3.1.2",
    "python-dotenv==1.2.1",
    "google-cloud-aiplatform>=1.137.0",
    "pydantic>=2.12.5",
    "gunicorn>=21.2.0",
]
